♻️ Extract core domain model into src/domain (Task_1)
Move api/types/domain.rs to src/domain.rs with sibling tests, absorb
internal/schema.rs as domain/schema.rs. api::types re-exports keep all
existing public paths working. Mechanical move, no behavior changes.
