+++ src/internal.rs
@@ -2,4 +2,3 @@ pub(crate) mod config;
 pub(crate) mod infrastructures;
 pub(crate) mod models;
 pub(crate) mod repositories;
-pub(crate) mod schema;
